chore: refactor browser composable
