+++ packages/ui/client/pages/index.vue
@@ -3,9 +3,9 @@
 import { Pane, Splitpanes } from 'splitpanes'
 import { browserState } from '~/composables/client';
 import { coverageUrl, coverageVisible, initializeNavigation, detailSizes } from '~/composables/navigation'
-import { provideResizing } from '~/composables/browser'
+import { useNotifyResizing } from '~/composables/browser'
 
-const { notifyResizing } = provideResizing()
+const { notifyResizing } = useNotifyResizing()
 const dashboardVisible = initializeNavigation()
 
 const mainSizes = useLocalStorage<[left: number, right: number]>('vitest-ui_splitpanes-mainSizes', [33, 67], {
